requirements.txt
# Core Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=0.19.0

# Web Dashboard Dependencies
flask>=2.0.0
flask-cors>=3.0.0

# HTTP Requests and API Integration
requests>=2.28.0
aiohttp>=3.8.0

# Text Processing and Fuzzy Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

# Database Support
aiosqlite>=0.17.0

# Production Web Server
gunicorn>=20.1.0

# Utility Libraries
certifi>=2017.4.17
charset-normalizer>=3.0.0
idna>=3.0
urllib3>=1.21.1

# Web Framework Dependencies
blinker>=1.9.0
click>=8.1.3
itsdangerous>=2.2.0
jinja2>=3.1.2
markupsafe>=2.1.1
werkzeug>=3.1.0